feat(GDB): self-contained init — move gun setup from Construct into GDB.js
GDB.js now owns its own initialization:
- reads Statics.site at eval time (safe: only loaded via dynamic import
  from Construct.GDB which runs after Construct.Site)
- creates gun instance with peers + OPFS from Statics
- exports gun (real instance), GUN (constructor), SEA
- sets globalThis.gun / globalThis.Gun / globalThis.SEA / globalThis.sea

SEA fallback: globalThis.SEA || globalThis.Gun?.SEA because sea.js
attaches to Gun.SEA, not a standalone global.

Construct.GDB simplified to a single await import() + guard checks.
Callers can now import { gun, GUN, SEA } from '/core/GDB.js' directly.
